Unificar modelado de Donaciones y Donantes
