+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/Bien.java
@@ -1,0 +1,3 @@
+package ar.edu.utn.frba.ddsi.donaciones.models.entities;
+
+public class Bien {}
