--- donaciones-service/src/main/java/donacion/DonacionIndependiente.java
@@ -1,22 +1,0 @@
-package donacion;
-
-import java.util.ArrayList;
-import java.util.List;
-import lombok.Data;
-
-@Data
-public class DonacionIndependiente {
-  private String descripcion;
-  private List<ItemDonacion> bienes = new ArrayList<>();
-
-  public void agregarBien(Bien bien, int cantidad) {
-    ItemDonacion item = new ItemDonacion();
-    item.setBien(bien);
-    item.setCantidad(cantidad);
-    this.bienes.add(item);
-  }
-
-  public void quitarBien(ItemDonacion bien) {
-    bienes.remove(bien);
-  }
-}
