--- donaciones-service/src/main/java/donacion/ItemDonacion.java
@@ -1,9 +1,0 @@
-package donacion;
-
-import lombok.Data;
-
-@Data
-public class ItemDonacion {
-  private Bien bien;
-  private int cantidad;
-}
